cleaned up routes
