+++ .nodemon.json
@@ -1,6 +1,5 @@
 {
   "exec": "npm-run-all --parallel generate-introspection start:development",
-  "verbose": true,
   "ignore": ["test/*", "client/*", "dist/*", "plugins/*/client"],
   "ext": "js,json,graphql,yml",
   "watch": [
