+++ routes/admin/index.js
@@ -1,6 +1,12 @@
 const express = require('express');
 const router = express.Router();
 
+if (process.env.NODE_ENV !== 'production') {
+  router.get('/docs', (req, res) => {
+    res.render('admin/docs');
+  });
+}
+
 router.get('/confirm-email', (req, res) => {
   res.render('admin/confirm-email');
 });
