+++ routes/api/index.js
@@ -1,17 +1,7 @@
 const express = require('express');
-const pkg = require('../../package.json');
-
 const router = express.Router();
 
-router.get('/', (req, res) => {
-  res.json({ version: pkg.version });
-});
-
-router.use('/assets', require('./assets'));
-router.use('/settings', require('./settings'));
-router.use('/auth', require('./auth'));
-router.use('/users', require('./users'));
-router.use('/account', require('./account'));
-router.use('/setup', require('./setup'));
+// Return the current version.
+router.use('/v1', require('./v1'));
 
 module.exports = router;
