--- routes/api/settings/index.js
@@ -1,29 +1,0 @@
-const express = require('express');
-const SettingsService = require('../../../services/settings');
-const authorization = require('../../../middleware/authorization');
-
-const router = express.Router();
-
-router.get(
-  '/',
-  authorization.needed('ADMIN', 'MODERATOR'),
-  async (req, res, next) => {
-    try {
-      let settings = await SettingsService.retrieve();
-      res.json(settings);
-    } catch (e) {
-      return next(e);
-    }
-  }
-);
-
-router.put('/', authorization.needed('ADMIN'), async (req, res, next) => {
-  try {
-    await SettingsService.update(req.body);
-    res.status(204).end();
-  } catch (e) {
-    return next(e);
-  }
-});
-
-module.exports = router;
